backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
anthropic==0.7.5
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0
